requirements.txt
streamlit==1.29.0
pyyaml==6.0.1
watchdog
